Add admin view and controller

+++ Mentor Finder/src/main/java/com/abc/WebApp2/repository/LevelRepository.java
@@ -14,8 +14,6 @@ import org.springframework.stereotype.Repository;
  * @author User
  */
 @Repository
-public interface LevelRepository extends JpaRepository<Level, Long> {
-
+public interface LevelRepository extends JpaRepository<Level, Integer> {
    
-    
 }
